Only show my characters
