+++ src/plugins/package/index.js
