+++ src/utils/database.js
@@ -9,22 +9,49 @@ exports.newDB = ( name, defaultElement = { user: [] } ) => {
     db[name].defaults(defaultElement).write();
 }
 
-exports.isInside = async ( name, key, index, value ) => {
+const isInside = exports.isInside = async ( name, key, index, value ) => {
     return db[name].get(key).map(index).value().includes(value);
 };
 
-exports.get = async ( name, key, index ) => {
+const get = exports.get = async ( name, key, index ) => {
     return db[name].get(key).find(index).value();
 };
 
-exports.update = async ( name, key, index, data ) => {
+const update = exports.update = async ( name, key, index, data ) => {
     db[name].get(key).find(index).assign(data).write();
 };
 
-exports.push = async ( name, key, data ) => {
+const push = exports.push = async ( name, key, data ) => {
     db[name].get(key).push(data).write();
 };
 
-exports.set = async ( name, key, data ) => {
+const set = exports.set = async ( name, key, data ) => {
     db[name].set(key, data).write();
 }
+
+exports.getID = async ( msg, userID ) => {
+    let id = msg.match(/\d+/g);
+    let errInfo = '';
+
+    if (msg.includes('CQ:at')) {
+        let atID = parseInt(id[0]);
+        if (await isInside('map', 'user', 'userID', atID)) {
+            return (await get('map', 'user', { userID: atID })).mhyID;
+        } else {
+            errInfo = "用户 " + atID + " 暂未绑定米游社通行证";
+        }
+    } else if (id !== null) {
+        if (id.length > 1) {
+            errInfo = "输入通行证不合法";
+        } else {
+            return parseInt(id[0]);
+        }
+    } else if (await isInside('map', 'user', 'userID', userID)) {
+        return (await get('map', 'user', { userID })).mhyID;
+    } else {
+        errInfo = "您还未绑定米游社通行证，请使用 【绑定】来关联米游社通行证";
+    }
+
+    return errInfo;
+};
+
